python/requirements.txt
fastapi
uvicorn
python-dotenv
openai
anthropic
google-generativeai
groq
pydantic